Move siphash.h to public
