+++ src/ds/test/hash.cpp
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #include "../hash.h"
 
-#include "../siphash.h"
+#include "ccf/ds/siphash.h"
 #include "siphash_known_hashes.h"
 
 #include <doctest/doctest.h>
